Add Pandas Challenge
Note I put the analysis at the top of the jupyter notebook.  I can't get the rubric to come up, but in looking through the instructions I don't think we are asked to write this to a file and so the notebook is all you need.  Please let me know if it is otherwise and I am happy to have it print to a file.  Thanks.